moving app deployer from event processing to analytics
