+++ components/event-publisher/org.wso2.carbon.event.publisher.core/pom.xml
@@ -38,8 +38,8 @@
             <artifactId>org.wso2.carbon.core</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.carbon.event-processing</groupId>
-            <artifactId>org.wso2.carbon.event.processing.application.deployer</artifactId>
+            <groupId>org.wso2.carbon.analytics-common</groupId>
+            <artifactId>org.wso2.carbon.event.application.deployer</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon</groupId>
