+++ components/event-receiver/org.wso2.carbon.event.receiver.core/pom.xml
@@ -36,8 +36,8 @@
             <artifactId>org.wso2.carbon.core</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.carbon.event-processing</groupId>
-            <artifactId>org.wso2.carbon.event.processing.application.deployer</artifactId>
+            <groupId>org.wso2.carbon.analytics-common</groupId>
+            <artifactId>org.wso2.carbon.event.application.deployer</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon</groupId>
